chore: add estrutura para testes iniciais
