+++ docker-compose.yml
@@ -1,0 +1,29 @@
+version: '3'
+
+services:
+    teste-postgres-compose:
+        image: postgres
+        environment:
+            POSTGRES_PASSWORD: "teste.james123!"
+        ports:
+        - "15432:5432"
+        volumes:
+        - D:\docker\postgres-james\data:/var/lib/postgresql/data 
+        networks:
+        - postgres-compose-network
+        
+    teste-pgadmin-compose:
+        image: dpage/pgadmin4
+        environment:
+            PGADMIN_DEFAULT_EMAIL: "adrianopons@gmail.com"
+            PGADMIN_DEFAULT_PASSWORD: "pgAdmin123!"
+        ports:
+        - "16543:80"
+        depends_on:
+        - teste-postgres-compose
+        networks:
+        - postgres-compose-network
+
+networks: 
+    postgres-compose-network:
+        driver: bridge
