+++ src/config/database.js
@@ -1,0 +1,14 @@
+module.exports = {
+  host: '127.0.0.1',
+  username: 'docker',
+  password: 'teste.james123!',
+  database: 'testeJames',
+  dialect: 'postgres',
+  operatorsAliases: false,
+  logging: false,
+  define: {
+    timestamp: true,
+    underscored: true,
+    underscoredAll: true
+  }
+}
